Checking in a glob.
A glob is a cluster of files,
packaged together by a reason.

In this case,
the reason is the pursuit of an accessible platform,
with a hint of mentoring in the comment blocks.

By continuing, you consent to the learning implied by deep understanding and artisanship.

We must evolve if we are to survive.